Rewrite of technical.md and starting xml-for-om tutorial
One thing left untouched here is the example of a CD symbol definition
(transc1#log). That needs more attention, because the implication given
is wrong (see issue OpenMath/CDs#15), and also it seems to be based on
an older version of the CD (but the error is also in the present CD).

In addition, I've included some OMS URIs such as
  http://www.openmath.org/cd/nums1#rational
which currently fail with messages such as
  The requested URL /var/www/www.openmath.org/www/nums1.xhtml
  was not found on this server.
I think that is a misconfiguration on (the old) www.openmath.org however,
so it probably shouldn't call for a change here.